II. SMPS
1. data-virtNoList.html, data-virtNoList.js, DataService.java
-> '가번호 배정 현황' 페이지 '수험생명' 행 제거

2. FormItemVo.java
-> 평가항목 양식에 항목 추가(출력물 제목 등)

3. ImageService.java, SystemService.java
-> '데이터 관리' 페이지에서 '삭제' 메뉴 제대로 동작하지 않던 것 수정

+++ src/main/java/com/humane/smps/service/ImageService.java
@@ -13,14 +13,16 @@ import java.io.InputStream;
 @Slf4j
 public class ImageService {
 
-    @Value("${path.image.examinee:C:/api/image/examinee}") String pathExaminee;
-    @Value("${path.image.univLogo:C:/api/image/univLogo}") String pathUnivLogo;
+    @Value("${path.image.examinee:C:/api/image/examinee}")
+    String pathExaminee;
+    @Value("${path.image.univLogo:C:/api/image/univLogo}")
+    String pathUnivLogo;
 
     public InputStream getExaminee(String fileName) {
         return getFile(pathExaminee, fileName);
     }
 
-    public InputStream getUnivLogo(String fileName){
+    public InputStream getUnivLogo(String fileName) {
         return getFile(pathUnivLogo, fileName);
     }
 
@@ -42,9 +44,10 @@ public class ImageService {
         for (String p : path) {
             File examineePath = new File(p);
             File[] a = examineePath.listFiles();
-            for (File file : a) {
-                file.delete();
-            }
+            if (a != null)
+                for (File file : a) {
+                    file.delete();
+                }
         }
     }
 }
